+++ src/main/java/io/github/aiarchguard/platform/identity/CurrentActorProvider.java
@@ -1,0 +1,6 @@
+package io.github.aiarchguard.platform.identity;
+
+public interface CurrentActorProvider {
+
+    CurrentActor currentActor();
+}